requirements.txt
pandas>=1.5.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0
google-auth>=2.0.0